requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
astroid==1.4.4
Babel==2.2.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
celery==3.1.20
cement==2.6.2
certifi==2015.11.20.1
colorama==0.3.6
cssmin==0.2.0
cssselect==0.9.1
decorator==4.0.6
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.1
django-angular==0.7.15
django-crispy-forms==1.6.0
django-filter==0.12.0
django-pipeline==1.6.5
django-pipeline-browserify==0.2.0
django-toolbelt==0.0.1
djangorestframework==3.3.2
docutils==0.12
dodgy==0.1.9
enum34==1.1.2
Flask==0.10.1
flower==0.8.4
frosted==1.4.1
futures==3.0.4
gnureadline==6.3.3
gunicorn==19.4.5
importlib==1.0.3
ipaddress==1.0.16
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.33
lazy-object-proxy==1.2.1
logilab-common==1.1.0
lxml==3.5.0
Markdown==2.6.5
MarkupSafe==0.23
mccabe==0.3.1
path.py==8.1.2
pep257==0.7.0
pep8==1.7.0
pep8-naming==0.3.3
pexpect==4.0.1
pickleshare==0.5
pies==2.6.7
pies2overrides==2.6.7
ply==3.8
prospector==0.11.7
psutil==3.4.2
psycopg2==2.6.1
ptyprocess==0.5
PyExecJS==1.1.0
pyflakes==1.0.0
pylibmc==1.5.0
pylint==1.5.4
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.1
pylint-flask==0.1
pylint-plugin-utils==0.2.3
PyReact==0.6.0
pyroma==1.8.3
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.7
PyYAML==3.11
redis==2.10.5
requests==2.9.1
requirements-detector==0.4.1
selenium==2.49.2
setoptconf==0.2.0
simplegeneric==0.8.1
six==1.10.0
slimit==0.8.1
static3==0.6.1
tornado==4.2
traitlets==4.0.0
vulture==0.8.1
Werkzeug==0.11.2
wheel==0.24.0
wrapt==1.10.6
xlrd==0.9.4
xlutils==1.7.1
xlwt==1.0.0
